Refactor MyPetFragment using ViewPager2, Add PetManagerFragment
