+++ client/android/app/src/main/java/com/sju18001/petmanagement/ui/myPet/MyPetActivity.kt
@@ -2,9 +2,8 @@ package com.sju18001.petmanagement.ui.myPet
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
-import androidx.fragment.app.Fragment
 import com.sju18001.petmanagement.R
+import com.sju18001.petmanagement.ui.myPet.petManager.AddPetFragment
 
 class MyPetActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
