+++ client/android/app/src/main/res/layout/fragment_pet_manager.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:layout_marginBottom="@dimen/nav_height"
+    tools:context=".ui.myPet.petManager.PetManagerFragment">
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/my_pet_list_recyclerView"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:clipToPadding="false"/>
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/add_pet_fab"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentEnd="true"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="end|bottom"
+        app:backgroundTint="@color/carrot"
+        app:tint="@color/white"
+        android:src="@drawable/ic_baseline_add_24"
+        android:layout_margin="16dp"/>
+
+</RelativeLayout>
